+++ debug.c
@@ -5,19 +5,19 @@
 void disassembleChunk(Chunk* chunk, const char* name) {
     printf("== %s ==\n", name);
 
-    for (int16_t offset = 0; offset < chunk->count;) {
+    for (int32_t offset = 0; offset < chunk->count;) {
         // after disassembling the instruction at the given offset,
         // it return the offset of the next instruction
         offset = disassembleInstruction(chunk, offset);
     }
 }
 
-static int16_t simpleInstruction(const char* name, int16_t offset) {
+static int32_t simpleInstruction(const char* name, int32_t offset) {
     printf("%s\n", name);
     return offset + 1;
 }
 
-int16_t disassembleInstruction(Chunk* chunk, int offset) {
+int32_t disassembleInstruction(Chunk* chunk, int32_t offset) {
     printf("%04d ", offset);
 
     uint8_t instruction = chunk->code[offset];
